requirements.txt
python-nmap>=0.7.1
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.2
colorama>=0.4.6
